Issue #604: copy some functions from labels.py to utils.py and delete the unused ones
