+++ deepreg/model/loss/label.py
@@ -4,39 +4,15 @@ from typing import List, Optional
 
 import tensorflow as tf
 
-from deepreg.model.loss.util import NegativeLossMixin
+from deepreg.model.loss.util import (
+    NegativeLossMixin,
+    cauchy_kernel1d,
+    gaussian_kernel1d,
+    separable_filter,
+)
 from deepreg.registry import REGISTRY
 
 EPS = tf.keras.backend.epsilon()
-
-
-def gaussian_kernel1d(sigma: int) -> tf.Tensor:
-    """
-    Calculate a gaussian kernel.
-
-    :param sigma: number defining standard deviation for
-                  gaussian kernel.
-    :return: shape = (dim, ) or ()
-    """
-    assert sigma > 0
-    tail = int(sigma * 3)
-    k = tf.exp([-0.5 * x ** 2 / sigma ** 2 for x in range(-tail, tail + 1)])
-    k = k / tf.reduce_sum(k)
-    return k
-
-
-def cauchy_kernel1d(sigma: int) -> tf.Tensor:
-    """
-    Approximating cauchy kernel in 1d.
-
-    :param sigma: int, defining standard deviation of kernel.
-    :return: shape = (dim, ) or ()
-    """
-    assert sigma > 0
-    tail = int(sigma * 5)
-    k = tf.math.reciprocal([((x / sigma) ** 2 + 1) for x in range(-tail, tail + 1)])
-    k = k / tf.reduce_sum(k)
-    return k
 
 
 class MultiScaleLoss(tf.keras.losses.Loss):
@@ -94,8 +70,12 @@ class MultiScaleLoss(tf.keras.losses.Loss):
             else:
                 losses.append(
                     self._call(
-                        y_true=separable_filter(y_true, kernel_fn(s)),
-                        y_pred=separable_filter(y_pred, kernel_fn(s)),
+                        y_true=separable_filter(
+                            tf.expand_dims(y_true, axis=4), kernel_fn(s)
+                        )[..., 0],
+                        y_pred=separable_filter(
+                            tf.expand_dims(y_pred, axis=4), kernel_fn(s)
+                        )[..., 0],
                     )
                 )
         loss = tf.add_n(losses)
@@ -323,89 +303,3 @@ class JaccardIndex(MultiScaleLoss):
 @REGISTRY.register_loss(name="jaccard")
 class JaccardLoss(NegativeLossMixin, JaccardIndex):
     """Revert the sign of JaccardIndex."""
-
-
-def separable_filter(tensor: tf.Tensor, kernel: tf.Tensor) -> tf.Tensor:
-    """
-    Create a 3d separable filter.
-
-    Here `tf.nn.conv3d` accepts the `filters` argument of shape
-    (filter_depth, filter_height, filter_width, in_channels, out_channels),
-    where the first axis of `filters` is the depth not batch,
-    and the input to `tf.nn.conv3d` is of shape
-    (batch, in_depth, in_height, in_width, in_channels).
-
-    :param tensor: shape = (batch, dim1, dim2, dim3)
-    :param kernel: shape = (dim4,)
-    :return: shape = (batch, dim1, dim2, dim3)
-    """
-    strides = [1, 1, 1, 1, 1]
-    kernel = tf.cast(kernel, dtype=tensor.dtype)
-    tensor = tf.nn.conv3d(
-        tf.nn.conv3d(
-            tf.nn.conv3d(
-                tf.expand_dims(tensor, axis=4),
-                filters=tf.reshape(kernel, [-1, 1, 1, 1, 1]),
-                strides=strides,
-                padding="SAME",
-            ),
-            filters=tf.reshape(kernel, [1, -1, 1, 1, 1]),
-            strides=strides,
-            padding="SAME",
-        ),
-        filters=tf.reshape(kernel, [1, 1, -1, 1, 1]),
-        strides=strides,
-        padding="SAME",
-    )
-    return tensor[:, :, :, :, 0]
-
-
-def compute_centroid(mask: tf.Tensor, grid: tf.Tensor) -> tf.Tensor:
-    """
-    Calculate the centroid of the mask.
-
-    :param mask: shape = (batch, dim1, dim2, dim3)
-    :param grid: shape = (dim1, dim2, dim3, 3)
-    :return: shape = (batch, 3), batch of vectors denoting
-             location of centroids.
-    """
-    assert len(mask.shape) == 4
-    assert len(grid.shape) == 4
-    bool_mask = tf.expand_dims(
-        tf.cast(mask >= 0.5, dtype=tf.float32), axis=4
-    )  # (batch, dim1, dim2, dim3, 1)
-    masked_grid = bool_mask * tf.expand_dims(
-        grid, axis=0
-    )  # (batch, dim1, dim2, dim3, 3)
-    numerator = tf.reduce_sum(masked_grid, axis=[1, 2, 3])  # (batch, 3)
-    denominator = tf.reduce_sum(bool_mask, axis=[1, 2, 3])  # (batch, 1)
-    return (numerator + EPS) / (denominator + EPS)  # (batch, 3)
-
-
-def compute_centroid_distance(
-    y_true: tf.Tensor, y_pred: tf.Tensor, grid: tf.Tensor
-) -> tf.Tensor:
-    """
-    Calculate the L2-distance between two tensors' centroids.
-
-    :param y_true: tensor, shape = (batch, dim1, dim2, dim3)
-    :param y_pred: tensor, shape = (batch, dim1, dim2, dim3)
-    :param grid: tensor, shape = (dim1, dim2, dim3, 3)
-    :return: shape = (batch,)
-    """
-    centroid_1 = compute_centroid(mask=y_pred, grid=grid)  # (batch, 3)
-    centroid_2 = compute_centroid(mask=y_true, grid=grid)  # (batch, 3)
-    return tf.sqrt(tf.reduce_sum((centroid_1 - centroid_2) ** 2, axis=1))
-
-
-def foreground_proportion(y: tf.Tensor) -> tf.Tensor:
-    """
-    Calculate the percentage of foreground vs background per 3d volume.
-
-    :param y: shape = (batch, dim1, dim2, dim3), a 3D label tensor
-    :return: shape = (batch,)
-    """
-    y = tf.cast(y >= 0.5, dtype=tf.float32)
-    return tf.reduce_sum(y, axis=[1, 2, 3]) / tf.reduce_sum(
-        tf.ones_like(y), axis=[1, 2, 3]
-    )
