+++ deepreg/model/loss/util.py
@@ -23,3 +23,68 @@ class NegativeLossMixin(tf.keras.losses.Loss):
         :return: negated loss.
         """
         return -super().call(y_true=y_true, y_pred=y_pred)
+
+
+def gaussian_kernel1d(sigma: int) -> tf.Tensor:
+    """
+    Calculate a gaussian kernel.
+
+    :param sigma: number defining standard deviation for
+                  gaussian kernel.
+    :return: shape = (dim, ) or ()
+    """
+    assert sigma > 0
+    tail = int(sigma * 3)
+    k = tf.exp([-0.5 * x ** 2 / sigma ** 2 for x in range(-tail, tail + 1)])
+    k = k / tf.reduce_sum(k)
+    return k
+
+
+def cauchy_kernel1d(sigma: int) -> tf.Tensor:
+    """
+    Approximating cauchy kernel in 1d.
+
+    :param sigma: int, defining standard deviation of kernel.
+    :return: shape = (dim, ) or ()
+    """
+    assert sigma > 0
+    tail = int(sigma * 5)
+    k = tf.math.reciprocal([((x / sigma) ** 2 + 1) for x in range(-tail, tail + 1)])
+    k = k / tf.reduce_sum(k)
+    return k
+
+
+def separable_filter(tensor: tf.Tensor, kernel: tf.Tensor) -> tf.Tensor:
+    """
+    Create a 3d separable filter.
+
+    Here `tf.nn.conv3d` accepts the `filters` argument of shape
+    (filter_depth, filter_height, filter_width, in_channels, out_channels),
+    where the first axis of `filters` is the depth not batch,
+    and the input to `tf.nn.conv3d` is of shape
+    (batch, in_depth, in_height, in_width, in_channels).
+
+    :param tensor: shape = (batch, dim1, dim2, dim3, 1)
+    :param kernel: shape = (dim4,)
+    :return: shape = (batch, dim1, dim2, dim3, 1)
+    """
+    strides = [1, 1, 1, 1, 1]
+    kernel = tf.cast(kernel, dtype=tensor.dtype)
+
+    tensor = tf.nn.conv3d(
+        tf.nn.conv3d(
+            tf.nn.conv3d(
+                tensor,
+                filters=tf.reshape(kernel, [-1, 1, 1, 1, 1]),
+                strides=strides,
+                padding="SAME",
+            ),
+            filters=tf.reshape(kernel, [1, -1, 1, 1, 1]),
+            strides=strides,
+            padding="SAME",
+        ),
+        filters=tf.reshape(kernel, [1, 1, -1, 1, 1]),
+        strides=strides,
+        padding="SAME",
+    )
+    return tensor
